geosim/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "geosim"
version = "0.1.0"
description = "Physics-realistic geophysical simulation engine for HIRT and Pathfinder instruments"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "pyzmq>=25.0",
]

[project.optional-dependencies]
viz = [
    "pyvista>=0.42",
    "jupyter",
    "jupyterlab",
    "matplotlib",
    "trame",
]
hirt = [
    "SimPEG>=0.20",
    "pygimli>=1.4",
]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "pandas",
    "ruff",
]
all = ["geosim[viz,hirt,dev]"]

[project.scripts]
geosim-server = "geosim.server:main"

[tool.setuptools.packages.find]
include = ["geosim*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "simpeg: requires SimPEG (deselect with '-m \"not simpeg\"')",
    "pygimli: requires pyGIMLi (deselect with '-m \"not pygimli\"')",
    "slow: marks tests as slow-running",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
